[CHORE] restructured package
