+++ src/main/java/com/shion1305/ynufes/bodytemp2022/contoller/ProcessorManager.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2022 Shion Ichikawa All Rights Reserved.
  */
 
-package com.shion1305.ynufes.bodytemp2022;
+package com.shion1305.ynufes.bodytemp2022.contoller;
 
 import com.linecorp.bot.model.event.Event;
 import com.shion1305.ynufes.bodytemp2022.config.InstanceData;
