+++ src/main/java/com/shion1305/ynufes/bodytemp2022/contoller/ReminderSender.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2022 Shion Ichikawa All Rights Reserved.
  */
 
-package com.shion1305.ynufes.bodytemp2022;
+package com.shion1305.ynufes.bodytemp2022.contoller;
 
 import javax.servlet.ServletContextEvent;
 import javax.servlet.ServletContextListener;
