+++ web/index.jsp
@@ -1,7 +1,5 @@
-<%@ page import="com.shion1305.ynufes.bodytemp2022.ProcessorManager" %>
-<%@ page import="java.util.HashMap" %>
-<%@ page import="java.util.Map" %>
-<%@ page import="java.util.ArrayList" %><%--
+<%@ page import="java.util.ArrayList" %>
+<%@ page import="com.shion1305.ynufes.bodytemp2022.contoller.ProcessorManager" %><%--
   Created by IntelliJ IDEA.
   User: shion
   Date: 2022/05/29
